[Read desc] Change bomb timer behavior in SuiseiAmongUs
Changed the bomb timer behavior to more accurately represent the gameplay time left instead of it including the "cutscene time" for when you lose. The cutscene is played after the timer ends, and the minigame ends right after.
The minigame is ended immediately on time over when won, since it'd be stupid to waste time staring at a cutscene not happening. Especially since the timer has already ended by then.